+++ tools/nixfmt.nix
@@ -23,5 +23,5 @@ nix-hs-utils.mkShellApp {
     # shellcheck disable=SC2046
     nixfmt "''${args[@]}" $(fd "$dir" -e nix)
   '';
-  runtimeInputs = [ pkgs.fd pkgs.nixfmt ];
+  runtimeInputs = [ pkgs.fd pkgs.nixfmt-rfc-style ];
 }