+++ flake.nix
@@ -7,13 +7,12 @@
     flake-parts.lib.mkFlake { inherit inputs; } {
       perSystem = { pkgs, ... }:
         let
-          ghcVers = "ghc962";
+          hlib = pkgs.haskell.lib;
+
+          ghcVers = "ghc981";
           compiler = pkgs.haskell.packages."${ghcVers}".override {
             overrides = final: prev: {
-              # For some reason, the cabal-fmt in nixpkgs does not have /bin
-              # i.e. no executable.
-              cabal-fmt = final.callHackage "cabal-fmt" "0.1.7" { };
-              implicit-hie = prev.implicit-hie_0_1_4_0;
+              stylish-haskell = prev.stylish-haskell_0_14_6_0;
             };
           };
 
@@ -47,14 +46,14 @@
             \tHaskell Miscellaneous:
             \t  - hie:         ${compiler.implicit-hie.version}
             \tNix Formatters:
-            \t  - nixfmt:      ${pkgs.nixfmt.version}
+            \t  - nixfmt:      ${pkgs.nixfmt-rfc-style.version}
             \t  - nixpkgs-fmt: ${pkgs.nixpkgs-fmt.version}
             \tInformation:
             \t  - help
             \t  - version
             See github.com/tbidne/nix-hs-tools#readme.
           '';
-          version = "0.9.1.0";
+          version = "0.10.0.0";
         in {
           apps = {
             cabal-fmt = import ./tools/cabal-fmt.nix compilerPkgs;